fetch_btc_ohlcv/requirements.txt
# crypto-indicator-service-docker/requirements.txt
google-cloud-bigquery
ccxt
pandas
ta
Flask
gunicorn
numpy